sql/row: move the low-level FK existence checker to its own file
Release note: None
